"Added sorting feature to Scores collection in ViewModel.cs"
Implemented a new feature in the `ViewModel.cs` file within the `RushHour2` namespace to sort the `Scores` collection. Now, whenever a new score is added to the `Scores` collection, it is sorted in descending order based on the `Points` property of each `Score` object using the `OrderByDescending` LINQ method. The sorted collection is then reassigned to the `Scores` property, ensuring the `Scores` collection is always sorted when a new score is added.